test_#2

+++ src/main/java/com/example/CryptoCurrencyWatcher/model/Currency.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import javax.persistence.*;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -25,12 +27,12 @@ public class Currency {
     private Double percentChange24h;
     private Double percentChange1h;
     private Double percentChange7d;
-    private Double marketCapUsd;
-    private Double volume24;
-    private Double volume24Native;
-    private Double csupply;
+    private BigDecimal marketCapUsd;
+    private BigDecimal volume24;
+    private BigDecimal volume24Native;
+    private BigDecimal csupply;
     private Double priceBtc;
-    private Integer tsupply;
-    private Integer msupply;
+    private BigInteger tsupply;
+    private BigInteger msupply;
 
 }
